// src/types.ts
export type Inbox = {
  id: string;
  object: 'inbox';
  name: string;
  description?: string | null;
  slug?: string | null;
  account_id: string;
  is_primary?: boolean;
  is_private?: boolean;
  created_at: string;
  updated_at: string;
};
// ...
export type WidgetIconVariant = 'outlined' | 'filled';

export type WidgetSettings = {
  id: string | null;
  inbox_id?: string | null;
  acccount_id?: string | null;
  title?: string;
  subtitle?: string;
  color?: string;
  greeting?: string;
  away_message?: string;
  new_message_placeholder?: string;
  show_agent_availability?: boolean;
  agent_available_text?: string;
  agent_unavailable_text?: string;
  require_email_upfront?: boolean;
  is_open_by_default?: boolean;
  custom_icon_url?: string;
  iframe_url_override?: string;
  icon_variant?: WidgetIconVariant;
  email_input_placeholder?: string;
  new_messages_notification_text?: string;
  base_url?: string;
  host?: string;
  pathname?: string;
  last_seen_at: string | null;
  account_id: string;
  created_at: string | null;
  updated_at: string | null;
  knowledge_base_settings: KnowledgeBaseSettings;
};

export interface KnowledgeBaseSettings {
  collections: {
    collectionId: string;
  }[];
}